requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
weaviate-client
python-multipart # For file uploads